+++ fliq/query.py
@@ -559,7 +559,7 @@ class Query(Generic[T], Iterable[T]):
         items: Iterable[Tuple[T, U]] = zip_func(self._items, *iterables)
         return self._self(items)
 
-    def slide(self, window: int, overlap: int, pad: Optional[T] = None) -> 'Query[Tuple[T, ...]]':
+    def slide(self, window: int, overlap: int, pad: Optional[T] = None) -> Query[Tuple[T, ...]]:
         """
         Yields a sliding window over the iterable. In practice, these are tuples of size 'window'
          containing the current element and the next 'size-1' elements
@@ -605,6 +605,23 @@ class Query(Generic[T], Iterable[T]):
             yield tuple(deque_window)
 
         return self._self(_window(self._items))  # type: ignore
+
+    def pairwise(self, pad: Optional[T] = None) -> Query[Tuple[T, T]]:
+        """
+        Yields tuples of consecutive elements in the query.
+        Practically this is a sliding window of size 2, with overlap 0.
+
+        Args:
+            pad: The value to use for padding the last window in case the iterable size is odd.
+
+        Examples:
+            >>> from fliq import q
+            >>> q([1, 2, 3, 4]).pairwise().to_list()
+            [(1, 2), (3, 4)]
+            >>> q([1, 2, 3, 4, 5]).pairwise().to_list()
+            [(1, 2), (3, 4), (5, None)]
+        """
+        return self.slide(window=2, overlap=0, pad=pad)
 
     def append(self, *single_items: T) -> Query[T]:
         """
@@ -848,6 +865,8 @@ class Query(Generic[T], Iterable[T]):
         non_sentinel_items: Iterable[Union[T, U]] = filter(lambda x: x is not sentinel,
                                                            flattened_items)
         return self._self(non_sentinel_items)
+
+
 
     # endregion