+++ docs/reference/release_notes.md
@@ -4,6 +4,8 @@
 
 * 🌟 **Slide** - new [slide method](code_api/mapper_methods.md#fliq.query.Query.slide) for creating
 tuples that "slide" over a query, in a windowed fashion (a sliding window).
+* 🌟 **Pairwise** - new [pairwise method](code_api/mapper_methods.md#fliq.query.Query.pairwise) for creating
+tuples that "slide" over a query, in a pairwise fashion (a sliding window of size 2, without overlap).
 
 ## v1.11.0
 * ⬆️ **Update to peek** - added fillvalue to the [peek method](code_api/peeking.md)